short unsigned int -> unsigned int. Thanks to the monster-buckets, any item-size can be handled.

+++ plugins/languages/cpp/parser/rpp/pp-macro.cpp
@@ -25,7 +25,7 @@
 
 using namespace rpp;
 
-short unsigned int pp_macro::itemSize() const {
+unsigned int pp_macro::itemSize() const {
   return constantSize(this);
 }
